requirements.txt
dash
geojson
loguru
pandas 
plotly
pyyaml
requests
